refactor(audio): fold WebAudioContextDiagnostics into AudioDiagnosticsUI.cs
Last standalone diagnostics record moves next to the others; AudioDiagnostics/
is now a single file.
